fix loop import issue
