+++ forum_engine/template.py
@@ -1,0 +1,16 @@
+class Template(object):
+	TEMPLATE = ""
+	NAME = ""
+	DB_SERVER_ADDRESS = ""
+	DB_SERVER_NAME = ""
+	CRAWLER_NAME = ""
+	LINK_TO_CRAWL = ""
+	COUNTRY = ""
+	THREAD_XPATH = ""
+	THREAD_LINK_XPATH = ""
+	LAST_PAGE_XPATH = ""
+	PREV_XPATH = ""
+	POST_XPATH = ""
+	FIELDS = dict()
+	CONDITIONS = dict()
+#end class
