+++ templates/crawler.arct
@@ -1,4 +1,4 @@
-from $ENGINE import Engine
+from $ENGINE.engine import Engine
 from $ENGINE import exceptions
 from pymongo import MongoClient
 import pymongo
@@ -144,5 +144,9 @@ class Crawler(object):
 				pass
 			#end try
 		#end for
+	#end def
+
+	def get_engine_info(self):
+		return engine.get_info()
 	#end def
 #end class
